move hook for reuse
